# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.CategoryCreate:
    properties:
      description:
        type: string
      name:
        type: string
      parent_id:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
  entity.CategoryUpdate:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      parent_id:
        type: string
      slug:
        type: string
    type: object
  entity.CommentCreate:
    properties:
      article_id:
        type: string
      content:
        type: string
      parent_id:
        type: string
    required:
    - article_id
    - content
    type: object
...
  entity.TagCreate:
    properties:
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
  entity.TagUpdate:
    properties:
      name:
        type: string
      slug:
        type: string
    type: object
...
  response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@news-api.dev
    name: News API Support
  description: REST API untuk website berita. Mendukung auth JWT, artikel, kategori,
    tag, dan komentar.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: News API
  version: "1.0"
paths:
...
  /articles/{id}/comments:
    get:
      description: Menampilkan komentar pada sebuah artikel (publik)
      parameters:
      - description: ID artikel
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Daftar komentar
          schema:
            $ref: '#/definitions/response.Response'
      summary: List komentar artikel
      tags:
      - Comments
...
  /categories:
    get:
      description: Menampilkan daftar kategori dengan pagination
      parameters:
      - default: 1
        description: Nomor halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Jumlah per halaman
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Daftar kategori dengan meta pagination
          schema:
            $ref: '#/definitions/response.Response'
      summary: List kategori
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: 'Membuat kategori baru (role: editor/admin)'
      parameters:
      - description: Data kategori
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CategoryCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Kategori berhasil dibuat
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Request tidak valid
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Buat kategori
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: 'Menghapus kategori (role: editor/admin)'
      parameters:
      - description: ID kategori
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Kategori berhasil dihapus
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Kategori tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus kategori
      tags:
      - Categories
    get:
      description: Mengambil detail kategori berdasarkan ID
      parameters:
      - description: ID kategori
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data kategori
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Kategori tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      summary: Detail kategori by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: 'Mengubah data kategori (role: editor/admin)'
      parameters:
      - description: ID kategori
        in: path
        name: id
        required: true
        type: string
      - description: Data yang diubah
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CategoryUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Kategori berhasil diubah
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Kategori tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update kategori
      tags:
      - Categories
  /categories/slug/{slug}:
    get:
      description: Mengambil detail kategori berdasarkan slug
      parameters:
      - description: Slug kategori
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data kategori
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Kategori tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      summary: Detail kategori by slug
      tags:
      - Categories
  /comments:
    post:
      consumes:
      - application/json
      description: Membuat komentar pada artikel (perlu login)
      parameters:
      - description: Data komentar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CommentCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Komentar berhasil dibuat
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Request tidak valid
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Artikel tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Buat komentar
      tags:
      - Comments
  /comments/{id}:
    delete:
      description: Menghapus komentar (perlu login)
      parameters:
      - description: ID komentar
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Komentar berhasil dihapus
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Komentar tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus komentar
      tags:
      - Comments
...
  /tags:
    get:
      description: Menampilkan semua tag
      produces:
      - application/json
      responses:
        "200":
          description: Daftar tag
          schema:
            $ref: '#/definitions/response.Response'
      summary: List tag
      tags:
      - Tags
    post:
      consumes:
      - application/json
      description: 'Membuat tag baru (role: editor/admin)'
      parameters:
      - description: Data tag
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.TagCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Tag berhasil dibuat
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Request tidak valid
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Buat tag
      tags:
      - Tags
  /tags/{id}:
    delete:
      description: 'Menghapus tag (role: editor/admin)'
      parameters:
      - description: ID tag
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tag berhasil dihapus
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Tag tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus tag
      tags:
      - Tags
    put:
      consumes:
      - application/json
      description: 'Mengubah data tag (role: editor/admin)'
      parameters:
      - description: ID tag
        in: path
        name: id
        required: true
        type: string
      - description: Data yang diubah
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.TagUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Tag berhasil diubah
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Tag tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update tag
      tags:
      - Tags
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"